Split off migration engine CLI from the core
Two reasons:

- compile times
- testing
