+++ migration-engine/core/Cargo.toml
@@ -7,35 +7,24 @@ edition = "2018"
 [dependencies]
 migration-connector = { path = "../connectors/migration-connector" }
 datamodel = { path = "../../libs/datamodel/core" }
-sql-migration-connector = { path = "../connectors/sql-migration-connector" }
+sql-migration-connector = { path = "../connectors/sql-migration-connector", optional = true }
 user-facing-errors = { path = "../../libs/user-facing-errors" }
-clap = "2.33"
-itertools = "0.8"
-
-chrono = { version = "0.4", features = ["serde"] }
-url = "2.1.1"
-
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
 
 anyhow = "1.0.26"
-thiserror = "1.0.9"
-
-jsonrpc-core = "14.0"
-json-rpc-stdio = { path = "../../libs/json-rpc-stdio" }
-
-futures = { version = "0.3", features = [] }
-tokio = { version = "0.2", features = ["rt-threaded", "macros"] }
 async-trait = "0.1.17"
+chrono = { version = "0.4", features = ["serde"] }
+futures = { version = "0.3", features = [] }
+itertools = "0.8"
+json-rpc-stdio = { path = "../../libs/json-rpc-stdio" }
+jsonrpc-core = "14.0"
+quaint = { version = "0.1", optional = true }
+serde = { version = "1.0" }
+serde_json = "1.0"
+thiserror = "1.0.9"
 tracing = "0.1.10"
-tracing-subscriber = { version = "0.1.6", features = ["tracing-log"] }
 tracing-futures = "0.2.0"
-quaint = { version = "0.1" }
+url = "2.1.1"
 
-[dev-dependencies]
-tempfile = "3.1.0"
-test-setup = { path = "../../libs/test-setup" }
-
-[[bin]]
-name = "migration-engine"
-path = "src/main.rs"
+[features]
+default = ["sql"]
+sql = ["quaint", "sql-migration-connector"]
